do not remove out files during mockup

+++ launcher.R
@@ -109,8 +109,10 @@ for (s in solutions) { #s = solutions[1]
       log_run(s, t, d, action="start", batch=batch, nodename=nodename, mockup=mockup)
       # TODO SRC_GRP_LOCAL is groupby specific
       Sys.setenv("SRC_GRP_LOCAL"=this_run[, paste(data, format, sep=".")])
-      if (file.exists(out_file)) file.remove(out_file)
-      if (file.exists(err_file)) file.remove(err_file)
+      if (!mockup) {
+        if (file.exists(out_file)) file.remove(out_file)
+        if (file.exists(err_file)) file.remove(err_file)
+      }
       ext = file.ext(s)
       cmd = sprintf("./%s/%s-%s.%s > %s 2> %s", ns, t, ns, ext, out_file, err_file)
       venv = if (ext=="py") sprintf("source ./%s/py-%s/bin/activate && ", ns, ns) else ""
